Redesigned UpdateHealthbar & added custom controller class
Removed UpdateHealthbar interface from player controller, as the demo will not be have multiplayer.

Instead this logic will be handled by a c++ implemented Interface on CharacterBase.

Also made a custom Controller class, that the Blueprint Controller inherits from. Currently this is done to more easily access a function from c++ to Blueprint, and allows further such logic - should it be necessary.

-- Yes this should have been a different branch.