Started new architecture
I changed how effects will work.
This approach is more data-driven, so all effects will be an archetype, instead of it's own class.

Also extrapolated some functions and variables, over to their own subclasses, rather than having it all in a base class. This is done to avoid unnecessary overhead.